Revert "Keysound Merging: Cast int..."
This reverts commit 29c1a7fb5a5b10b1abf6243962c425d54d9082c3.

+++ TECHMANIA/Assets/Scripts/Components/ResourceLoader.cs
@@ -415,8 +415,8 @@ public class ResourceLoader : MonoBehaviour
                         {
                             channelIndex *=
                                 channelIndex == 0
-                                    ? (int) (-Math.Min(note.volumePercent, 0) / 100d)
-                                    : (int) (Math.Max(note.volumePercent, 0) / 100d);
+                                    ? -Math.Min(note.volumePercent, 0) / 100d
+                                    : Math.Max(note.volumePercent, 0) / 100d;
                         }
                         mergedSamples[clipStartOffset + i] += (float) (
                             clipSamples.samples[clipIndex + channelIndex]
